New outputfile: polyA_....!
The output has been split up into two files; the utr_... file is poly(A)-read dependend
while the polyA_... file is poly(A)-read independent. In the first file
evidence for the length of the utr, independend of poly(A) reads, is given. In
the second, all the poly(A) sites in a utr is given, along with evidence for
this site.

On the programming side, I have made two new classes: one for each output file.
The utr that is being written has its own class which is used by the output
classes to calculate and write output. You did this because having just one
output class (like you had at first) became unmanagable; the class got too
big. The point of using classes for the output in the first place was to reduce
the amount of parameters passed in each function.

The new output caclulating classes have many function calls; you have realized
that writint the output takes a fraction of the time of the whole operation, so
there is no need to worry about the overhead. Better make it understandable than
worry about time.