Writing multi-exon 3UTRs to bed-file
Single-exon 3UTRs (whose gene only has single-exon 3UTRS) are treated
separately from the multi 3-UTR exons.

Single-exon 3UTRs are saved to the bedfile with a 1_1 after the transcript ID.
The first 1 indicates that this is exon number 1 from 3UTR number 1 for this
gene.

Multi-exon 3UTRs are saved in the same way. Typically you will see something
like 3_2, which tells you that this is exon 3 for the second utr of this gene.

Note that the multi-exon 'exons' are not native exons. They could have been
merged with other exons depending on their start location.

There is a controversy over whether to create N 3UTRs for each gene, depending
on UTR start position, or whether to make only 1 3UTR for each gene, where this
3UTR has merged all the possible exons for this gene.

For now the first option is implemented. The second should be considered.