Added 1 gene -> 1 utr for multi-exon utrs
Now it is possible to merge all exons for all 3UTRs for one gene. This way we
end up with 1 gene -> 1utr the same way that before it was 1 transcript -> 1
utr, which simplifies a lot of the work.

However, one inconsistency remains. For the genes with only one exon, it is
still possible to have 1gene -> several UTRs as long as the start sites of
those utrs don't cluster together.