Now limits 3UTRs to start from 3'-most CDS exon
Before UTRs were included in the set even if there were CDS exons overlapping
the UTR. Obviously, I can't disginguish expression of CDS exon from UTR exon.
Thus I implemented a function that checks each damn UTR for CDS exons from the
same gene. If they exist, the UTR is trimmed to start from the 3' end of the
CDS exon.

Sometimes this is bad, because the CDS exon can be a 200 nt piece in
the middle of a 3000 nt 3UTR -- effectivelty cutting that UTR in half.

A solution to this would be in those cases (when a CDS exon lands in the middle
of a long 3UTR) to trim the 3UTR only of the CDS exon is more than 30% of the
3UTR length, or if it is close to the end of the 3UTR.

The way it stands now I don't know if this is worth the effort. If I see that
this affects me, I can always go back to it. It would not be so difficult to
implemement.

In the coverage phase, you would have to be aware that you have kept some CDS
exons though. Maybe you could save them and use that information to remove
false positives.

Or maybe this is not worth the effort!! You don't have so much time!!