Changed rules for 3UTR-CDS filtering
Before 3UTS could be trimmed. After talking to Pedro, there is no more trimming
of 3UTRs.

For a 3UTR cluster, the individual 3UTRs in the cluster are checked for
internal CDS. If they contain a CDS, they are removed, except if the CDS is
less than 10% of the UTR exon it resides in. Then it is being ignored.

Finally, only one UTR is written per cluster. I simply choose the longest one.

The previous approach was too complicated and less conservative.