Change to utr filtering method
Now finally there is no more cutting of UTRs. The only UTRs that are included
are those that fall in introns or have a CDS less than 1% of the size UTR exon
in them.

This leaves open the problem of UTRs from other genes... next will be removing
these using BedTools.