Multi-exon 3UTRs now correctly assembled from annotation
The work up to writing the result is now done. The multi exons utrs are denoted
in the bedfiles with for example 'ENSG00000169598_2_1'. This means that the
entry is in gene ENSG00000169598, UTR number 2 from this gene, and it is exon
number 1 from this UTR. The 'value' field of the bedfile tells how many exons
in total this UTR has. For example
ENSG00000169598_2_1	3
means that this is the first exon out of 3 in UTR ENSG00000169598_2.

What remains now is to re-assemble the multi-exon utrs when writing to file.