Multi exon no saves to file correctly
The multi-exon effort is now over.

Changes are to the calculate and write of output.

The biggest effort was to collect the utrs with multiple exons. They would
often come in different parts of the coverage file.

My solution was to create a dictionary that collected the different utrs. When
the dictionary was full (len(exons) == tot_exons) I merge the exons and write
to file as normal.

What could still be done is adding to the output the begs and ends of the
exons.