New output file for tuning the cumulative cut-off
In order to justify the 99.5% value, I'm calculating the cut-off
for the polyA reads, when a polyA read is close to the 99.5% value.

If I iterate on this, I should be able to get to an optimal cut-off.

An output file 'cumul.stat' is put in the output folder if the 'tuning'
setting is set in the UTR_SETTINGS file. Note that polyA read must
be True or string for this to work.

Because of this, I have deleted the output in the polyA file that contained
similar values. In case I ever want it back, go to the commit before this
one and find it.