New feature: .bed to bigWig format conversion
Added the possiblity to format the polyA reads and the rest of the reads into
the bigWig format. See [BIGWIG] section in UTR_SETTINGS for details. After the
conversion a track line is printed. This track line can be used in the USCS
browser for easy browsing.

The bigWig format is indexed to hg19. The genome browser reads only those
portions from your file that it needs. The file is too big to upload to the
browser, so you keep it in a directory accessible from the outside so that the
browser can read from it.