Updated a lot of docstrings + two bugfixes
Updated a lot of docstrings for methods and classes.

Fixed a bug where the PAS sites were in a set; some of the code assumed the PAS
sites were ordered, and a set is not ordered. I changed set to list.

Fixed a bug where the output of the polyA site would be according to the
extended UTR. Since the output is now in terms of the un-extended UTR, this was
changed.