Plotting script up and running
The plotting script reads and parses the output from the 3UTR pipeline and
places it in datastructures (classes) for easy handling.

So far the boxplot, the scatterplot, and a triangluar scatterplot are
implemented. The latter shows the scatterplot of f.ex. 4vs4 variables at the
same time, and removes the redundant plots and the diagonal, forming a lower
triangular matrix plot.