Bugfix: internal exons wrongfully extended + new output file for genome browsing
A bug caused internal exons to have their beg_nonextended and end_nonextended
modified when in fact no extension had happened in the first place. The code
now only unextends exons that have been extended (namely the 3' ones).

As well, an output file of the 3UTR lengths, as output in the length file, is
now being provided together with the bigwig files. This file can be used with
a genome browser to easily visualize how long we predict our 3UTRs to be. Taken
together with the general coverage and the polyA coverage, this should
give a good visual help to debugging any remaining errors in the code.