Merged dataset-clusters in to mega_cluster + poly-coverage scatterplot
I merged all clusters into a mega-cluster. This way it is is easy to know if a
given cluster is part of a mega-cluster or not. An intermediate mapping dict
was made from dset_key -> mega_key.

Correlated the polyA coverages of those polyA clusters with more with
representation in more than one datset. Correlation coefficient around 0.5, not
very clear. It's not a bad thing: could mean different expression
levels in different compartments.