Fixed bug where the RPKM of multi-exon constructs was incorrectly added
RPKM is calculated as

 ((10**9)*(reads_in_UTR)/(total_nr_reads_in_experiment*length_of_utr)

My initialf formula for adding RPKMs was naive: it was a simple addition
weighted by the lengths. RPKMs cannot be calculated like that.

I needed to pass on the total number of reads variable into the method that
updates multi-exon-utrs, and then I needed to recalculate the reads of the old
and new UTRS, and then add them together.