BUGFIX: multi-exon joining did not include coverage vectors
There was a bug where the program tried to access a polyA site index of the
coverage vector in a multi-exon UTR, however that coverage vector
had not been extended when the UTRs were joined, so an index out of bound error
was reached.

This bug happened because when I merged the UTRs I thought that I would only
merge those features I thought would be useful. Lesson: merge everything, you
never know what you might need in the future.