Added aTTS distance to length-outout. Added various plots.
The epsilon_coord in the length-output file is now relative to the chromosomal
coordinat system, as expected.

The 'length' output file now outputs the distance from the epsilon-determined
length of the 3UTR to any annotated TTS site.

Added an outputfile that gives the TTS sites in the Gencode annotation that are
verified by our poly(A) reads.

Started on a method to give the false negative rate in results.py.