Added new plot: sensitivity of recovery of poly(A) clusters
Added a plot that shows that the sensitibity of recovery of poly(A) clusters is
high for 3UTRs with high RPKM, and low for 3UTRs with low rpkm -- as I had
hoped. This means that there is nothing mystical about our datasets -- more
expressed 3UTRs have a higher probability of being reported with poly(A)
clusters. Now, whenever I want to compare poly(A) usage between samples, I can
point to this distribution and say that I only want to consider UTRs with RPKM
over 30, or something like that.