Added DEBUGGING dir for debugging sessions
Now when the script is in DEBUGGING mode, the temp_output, output, and
source_bedfiles directories will be created inside a DEBUGGING directory. In
this way, you can run all the 'real' simulations with the default directories,
while you can continue working on the script in the debugging directory. I wish
I had done this in the beginning!