Two new plots: rpmk-dependence of eps_len and coverage up/downstream
The first plot shows that the 3UTRs are independent of RPKM in terms of length.
It also shows that, on the other hand, the relative epsilon length is sensitive
to RPKM, but only for low values < 1. Thus when fetching RPKM values, I think
we can trust the rel_eps_size for 3UTRs with RPKM < 1.

The second plot shows that the before/after coverage ratio is not a good
determiner of if we have a epsilon_end or not (at least not with
epsilon=0.999!). The stochasticity of coverage might ruin this variable.