Added gem-index-filepath to UTR_SETTINGS and updated docstring
If the Utail program should extract poly(A) reads from the gem-mapped files,
the gem mapper needs an index file. Actually, gem-do-index creates several
index files with several suffixes. In the UTR_SETTINGS file, in the
[POLYA_READS] section, you must supply the path to the index files, without the
suffix! So if the directory where the index files are are looks like this:

 470M Oct 18  2009 H.sapiens.genome.hg19.main.bwt.bl0
 470M Oct 18  2009 H.sapiens.genome.hg19.main.bwt.bl1
 470M Oct 18  2009 H.sapiens.genome.hg19.main.bwt.bl2
 470M Oct 18  2009 H.sapiens.genome.hg19.main.bwt.bl3
 1.8K Oct 18  2009 H.sapiens.genome.hg19.main.bwt.toc
 683M Oct 18  2009 H.sapiens.genome.hg19.main.csa
   18 Oct 18  2009 H.sapiens.genome.hg19.main.gem
 6.9K Oct 18  2009 H.sapiens.genome.hg19.main.loc
  17K Oct 18  2009 H.sapiens.genome.hg19.main.log

you must supply

"gem_index_file = somedir/H.sapiens.genome.hg19.main",

where somedir is the directory where the index files are kept.

As well, I updated several docstrings for methods and classe for the automatic
documentation.