Final bugfix on multiexon coverage vectors
Multi exon coverage vectors were assembled the wrong way for -strand.

The 'sequence' array is strand-sensitive. It is always in the 3->5 diretction,
so how you add the sequences from the multi-exons must take that into the
account.

However, the 'covr_vector' is not strand sensitive. You thought it was by
mistake. Now this is fixed.

Also added some minor safety-fixes when reading the UTR_SETTINGS file.