New cross-cellline, cross-compartment study of 3UTR lengths
I show that UTRs with different length between compartments are conserved
between cell lines. The number of 3UTRs with different 3UTR lenghts is much
higher than what would be expected by random sampling.

The 3UTRs do not have the same direction of length difference. This, however,
can not be explained by the RPKM variation between the compartments for the
different datasets. The explanation is unknown.