New basic polyA statistics and started with nucleosome counting
You rewrote the poly(A) read counter because Roderic wanted stats for comparing
poly(A)+ and poly(A)- samples.

You started on the nucleosome method. You realized that instead of importing
the coverage vector for the 3UTR object, it's better to export the poly(A)
sites in bed-files, extend them by 500 nt, and run intersectBed on them. This
way you can export whatever poly(A) sites you want to, and use the same
downstream code for treating them all.