Added 'get_length' to settings-file and made preparations
It seems that the original output writer is too dependent on reading from the
coverage files. Since there is no point in calculating the coverage file if you
just want poly(A) stats. Therefore you have started on a "only_polyA_writer"
method. The goal is to output the same information you would get with
"get_length = True". The plus-side of this story is that the process will be
twice as fast. The bottleneck will then definitely be getting the reads. Should
you possibly store the unmapped reads locally on your drive? You can just zcat
them out! 90% are mapped so the footprint would not be more than about 1GB per
dataset. You would need maybe 20GB in total. This would speed up your pipeline
considerably ... For example you could just save the raw GEM-line when you
detect that it is a poly(A) read. Then, whenever a program is run again with
this dataset, you could check if you have already gotten the poly(A) read. The
danger of course would be if you change dataset files but keep the same name
...