Now saving genome-region as part of the output file
Now the genome region (3UTR, noncoding exons ...) will be part of the filename
for the output file. The region is taken from the filename of the region, so
this filename must be in the format '3UTR_smth_smthlese.bed'. split('_')[0] in
other words.

Having the genome region as part of the filename will alow a comparison between
genomic regions.