Added 'region' to the dset_id parameter + added a poly(A) cache
The genomic region is now part of the output files. They can be 3UTR or
whatever, right. The naming convention is 3UTR_exonic for the region files.
Only intergenic is different, but I have a special case test for that.

Also I'm saving poly(A) read output in a poly(A) cache if this is asked for.
Saves the time to read through the original files to re-obtain the data.