Now making annotated 3UTRs if annotation is present
Now you make annotated 3UTR regions if the annotation is present. Optionally,
make this a .bedfile and don't re-run the annotation parser.

As well, you no longer re-calclulate the poly(A) reads for each region! omg! :)
you only need to do it once: after that, it's just a matter of intersecting the
known reads.

What you need to do now, is let the program know if it's dealing with reads
from an annotation or not. if it is, it must "move" the poly(A) reads from one
strand to another. If not, it should not move them.