Feature: now checking if strand is known or no
If the feature we're looking at is not in the annotation, the program now scans
both possible downstream sequences looking for the poly(A) site. However, this
could be considered a bug: if the data is stranded, you should look only on the
opposite strand of the strand which the poly(A) reads map to. This is known.