New settigns: pairedend and stranded
You need to know if the incoming reads are pair ended. Pair ended reads allow
you to determine the strand of the transcript from which the poly(A) reads
come.

When not using an annotation, you need to know if the strand of the genomic
regions in the supplied .bedfiles are known or not.