Updated p(A) strip algorithm + only getting genomic sequence when needed
Before this sequence would be lost to us XXXXXXAAAAAAAAAAAAGATAAAA, although
two misincorporations after one another is not unlikely. Now I capture these
reads. Before only 1 mismatch was allowed, now 2 are allowed. Also there are
paramters to control this. If the parameter is set to 5, then you need 4/5 As,
or 8/10. Setting this at 6 and the minimum tail length at 5 gives a good
balance, IMHP!

Now, for only poly(A) output, sequences are not obtained for the whole region
in question. That was a waste of memory. Now I only get the sequences that are
necessary, that is , 40 nt downstream a putative poly(A) site.

This is TODO for the utr-tail-length output.

You have become aware of a caveat: since you strip parts of a read and remap,
some of your remaps will be from exon-exon junctions. A possible solution to
this is to trim your final poly(A) reads with known exon-exon junctions, or
lettings the split-mapper have a go at the dataset and using subtractBed from
the splitmapper bedfiles.